requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
deep-translator==1.11.4
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-translate==3.13.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio_client==0.9.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
jeepney==0.8.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
langdetect==1.0.9
lxml==5.1.0
lz4==4.3.2
MarkupSafe==2.1.3
newspaper3k==0.2.8
nltk==3.8.1
openai==1.13.3
orjson==3.9.10
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.5.3
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
replicate==0.22.0
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
youtube-transcript-api==0.6.2
